+++ NewConsoleWriter.Tests/ConsoleCollection.cs
@@ -1,0 +1,4 @@
+﻿[CollectionDefinition("Console Collection", DisableParallelization = true)]
+public class ConsoleCollectionDefinition
+{
+}